refactor: split SlateTest.ts by element type (Option A)
The 2448-line SlateTest.ts is split into per-element test files that
mirror the src/elements/ layout:

  tests/PointTest.ts         22 tests
  tests/LineTest.ts          17 tests
  tests/CircleTest.ts         7 tests
  tests/PlaneTest.ts          4 tests
  tests/PolygonTest.ts       14 tests
  tests/SectorTest.ts         4 tests
  tests/SphereTest.ts         3 tests
  tests/PolyhedronTest.ts     4 tests
  tests/ColorsTest.ts        28 tests
  tests/ParseParamTest.ts    18 tests
  tests/SlateTest.ts         14 tests  (slate-level: construct/update/
                                        drag/init/error paths/closest)

Shared helpers extracted to:
  tests/shared/testHelpers.ts  — almostEqual, toElements
  tests/shared/dragScenes.ts   — pivotScene_data, scene3d_data,
                                 specialized_data + their buildXxxScene()
                                 wrappers used by slate drag tests

All 135 tests preserved verbatim — this is organizational only, not a
content change. npm run test:unit now globs these explicit files; the
snapshot suite is unchanged.
